Improve resolution of CAV patients via identifiers
This is temporary, because it's likely that we'll have a resolver that looks up based on organisation/identifier rather than hard-coding mutation. That will decouple client applications from having to know these details. But this is for the proof-of-concept, and of course the resolvers will work whatever mechanism generates the CAV namespaced patient data.