Add more of the referral workflow.
Now have a landing page that does user login and choice of patient. If either are already known, those steps can be autocompleted so it can work standalone or as part of the wider application. The most difficult thing is this UI work is not the data or the integration with the server, but fighting CSS styling. It is very difficult to do anything slightly different to what was originally designed by the framework.