Authenticate EQL operations.
I originally parsed the EQL and had a whitelist of operations that did not need authentication. I did not like it. It felt messy. Instead, we now have different endpoint for requests that do not need authentication. That means we can carefully curate the request for those entry points - and ensure we do not have boundless requests. All requests to the main API endpoint are now authenticated.