Add resolution of active projects for those users registered with rsdb.
If a user is registered to the legacy rsdb backend, we add resolvers in order to jump from NHS Wales user into rsdb and from there to a list of active projects. This enables us to provide different functionality for registered users while being able to expose some workflows for non-registered, but authenticated users.