Add function to generate drug regimen calculation.
This gets all medications and then groups by DMT, and then uses the 'from' date from the first entry in that batch, and the 'to' date from the last entry in that batch. This gives a good representation of when a treatment started and stopped - except for the following caveats: it won't capture gaps in treatment and it cannot deal with overlapping regimens.