Tidy up handling of DMTs, including mapping all products.
Unfortunately, we have AMPPs and VMPPs in the legacy dataset. These must be converted to AMPs or VMPs in order for categorisation, or we must include AMPPs and VMPPs in the codelists. I think this is a data quality issue, so next step will be to highlight all patients with a medication recorded as AMPP or VMPP so that they can be fixed.