Add view patient by project pseudonym route
Add a method to load a patient record using a tuple of project identifier and pseudonym, and fetch patient record as 'current patient' but in the context of the project. That means we can provide a project-specific prism of that record, while leveraging the same data, and mostly the same UI components.