Fix sequence generation
This is a real gotcha. Horizontal inheritance in the legacy WebObjects application uses a sequence in the table representing the abstract  superclass, and not the sequence generated in the table of the 'entity'. So, we need to force usage of the 'correct' (sic) sequence when creating a new multiple sclerosis specialist summary table.