Switch to API on same port
Port 8000 was causing issues for some firewalls, e.g. within the NHS. So now we reverse proxy certain URIs to the /api and /login endpoints provided by the server, and try to satisfy other requests from the filesystem. Here's the example Caddyfile to which this operates:

patientcare.app {
        root * /var/www
        reverse_proxy /api localhost:8080
        reverse_proxy /login localhost:8080
        file_server
}
www.patientcare.app {
        redir https://patientcare.app{uri}
}