Add initial commit of pc4-api
This provides a hopefully sane API for the modular pc4 functionality for java clients. We'll build an uberjar that contains all dependencies because that makes it easier to consume by the legacy rsdb build, which still uses ant! We could split these out and provide these in the target products as a convenience, making this redundant. To be decided.