Switch to using published artefacts; use new clods
As we need to make use of nhspd outside of clods, it is better to inject nhspd as a dependency into clods, rather than have clods manage nhspd itself. The new clods release supports this approach, but can still use the old approach in which it loads and manages the nhspd. This therefore creates a distinct integrant service for nhspd that is used in the initialisation of clods.