Tidy up development, add instruments and function specs
Attempt to make it very difficult to mix up the use of a patient primary key with a patient identifier. They are often the same, but not always due to legacy reasons. The basic rule is that outside of a few specialised areas, the use of a patient primary key is from within a map with an explicit key, :t_patient/id, while patient-identifier can be used more widely as a user visible identifier.