Add MRI annotations. Switch to re-frame for managing modal state
Permit complex T2/gad enhancing recording including comparison to prior imaging. Switch to using re-frame state to manage appearance of modal dialog for editing. This was using a local atom instead, but this is much better. It means we can prevent closure of modal dialogs if there are validation errors. It also provides a hook to pass back server-side validation messages.