Add custom ECL for anti-platelet
Unfortunately, not all products within the UK SNOMED CT drug extension have ATC codes in the UK dm+d product. For example, 'Aspirin' 7947003 is a product containing aspirin, which includes a range of products that contain aspirin. The ATC mapping as currently implemented does not include 7947003 because its ATC code in dm+d is "n/a". It's unclear to me whether this can be solved in a generic way at the moment, although it is easy to compose hermes and dmd in order to explore the relationships:

For example,

(set (->> (hermes/get-all-children (:com.eldrix/hermes system) 7947003)
          (map #(dmd/atc-for-product (:com.eldrix/dmd system) %))))
=> #{nil "N02BA01" "N02BE51" "N02AJ07" "B01AC06" "N02BA51" "n/a" "C07FX04"}

shows that the ECL expression <<7947003 will encompass products from multiple ATC codes. That is fine, as ATC includes support for combination products. The alternative is to use the ingredient models in the drug extension or the raw dm+d data. For the moment, we use the simplest option and simply add an ECL inclusion list to the research study's medication definition.

The future option would be to use the ATC code to expand to a list of products, take their ingredients and then use active ingredients for matching in codelists. It's likely the exact choices in codelists will need to be determined on a case-by-case basis, but it is clear that the is-a relationships don't always give expected results.