Improve integrant and pathom operation configuration
For library jar, we simply want pathom operations. Leave out HTTP implementations until needed. We define resolvers using our master configuration, and so some namespaces won't be loaded unless needed. This then means the library jar can be much smaller and more easily embedded into a larger application, while preserving the ability to continue to develop pc4-server as a standalone service. It remains unclear whether to continue to use the pedestal backend, or the fulcro backend - this will depend on the front-end choices, which I have not fully decided, as yet.