Remove batch resolving, for now
This didn't work, as there is no guarantee of order when using SQL 'in' clauses. Fortunately, pathom decided to ignore the batch as data were returned as a sequence and not a vector, but this issue manifested with the commit switching to generating vectors. For optimisation, it may be batch resolution would be a good idea, but this makes it function correctly.