Derive common concepts for a user
We do this in the same way as the legacy rsdb, by making use of a users' project membership to derive common concepts. However, this is much simpler, and takes into account only active projects and active memberships. We make use of the new intersect-ecl and so are able to constrain arbitrary sets of concept identifiers to a SNOMED ECL expression and therefore populate pop-up lists of common concepts in many different contexts just like the legacy rsdb system.