Add server-side generation
This remains an experiment, but is much more closely aligned to the old WebObjects rsdb version. Most HTML is server side generated, but we define ui components using cljc code which can be compiled on the JVM or converted into Javascript for ad-hoc client side HTML generation using React as and when necessary, or more old-fashioned DOM manipulation.