Patient matching in preparation for registration
Rename namespace to 'demographics' given this is all about matching and updating core demographics from an external auhority. Registration from an authoritative requires managing an existing matching patient based on the same identifiers (the easy path) and a match on basic demographics (the much more difficult path, only needed because of legacy data in which patients were registered manually in the multiple source systems that were migrated to PatientCare).