Fix HTML routing for patients
We have two top-level routes to access a patient record - one in the context of a patient, and one without. Internal routing within a single patient record uses only internal dynamic routing and so does not change the URL. As such, back and forward buttons work, and preserve context of 'current project' in the URL when appropriate.